fix: usize→u32 for DiscoveryResult.total_models (Specta BigInt on 64-bit)
Comprehensive sweep: no remaining u64/i64/usize in specta::Type structs
without #[specta(type = f64)] annotations.

+++ backend/src/inference/model_discovery/types.rs
@@ -114,7 +114,7 @@ pub struct DiscoveryResult {
     /// Per-provider results.
     pub providers: Vec<ProviderDiscoveryResult>,
     /// Total number of models discovered.
-    pub total_models: usize,
+    pub total_models: u32,
     /// Providers that had errors.
     pub errors: Vec<String>,
 }
